Enemy damaging, health, tower dot application

+++ core/src/win/idecm/towerdefence/EnemyKind.java
@@ -4,5 +4,5 @@ public interface EnemyKind {
     String getTextureName();
     double getSpeed();
     double getSize();
-    int getMaxHealth();
+    double getMaxHealth();
 }
